+++ Frontend/src/app/(pages)/admin/dashboard/page.tsx
@@ -3,6 +3,7 @@
 import withAuth from "@/app/_components/admin/Auth/withAuth"
 import ChatPanel from "@/app/_components/admin/chat/chat-panel"
 import Sidebar from "@/app/_components/admin/sidebar/sidebar"
+import Stats from "@/app/_components/admin/stats/stats"
 import WebSocketServiceContext from "@/app/_core/services/websocketService"
 import { useContext, useEffect, useState } from "react"
 
@@ -12,10 +13,8 @@ function Dashboard(){
 
 
     return(
-        <section className="bg-[#222] w-screen h-screen md:grid md:grid-cols-7 md:grid-rows-1 md:gap-3">
-            <Sidebar/>
-            
-            <ChatPanel/>
+        <section className="bg-[#222] w-screen h-screen p-10">
+            <Stats/>
 
         </section>
     )