+++ Frontend/src/app/_components/admin/chat/chat-panel.tsx
@@ -1,39 +1,47 @@
 import { useContext, useEffect, useState } from "react";
-import Chat from "./chat";
+
 import DataServiceContext from "@/app/_core/services/dataService";
 import httpRequestContext from "@/app/_core/services/httpRequest";
 import { IHttpOptions } from "@/app/_core/interfaces/core.interface";
 import { data } from "autoprefixer";
-import { IChat, IClient } from "./interfaces/chat.interface";
+import { IChat, IChatPanelProp, IClient } from "./interfaces/chat.interface";
 import { ChatStatusEnum } from "./enums/chat.enum";
 import { Loading } from "@/app/_core/components/loading/loading";
 import WebSocketServiceContext from "@/app/_core/services/websocketService";
 import { WebsocketTypeEnum } from "@/app/_core/enums/core.enum";
 
+import Chat from "@/app/_components/chat/chat"
+import { UserTypeEnum } from "../../chat/enums/chat.enum";
+import { IChatData } from "../../chat/interfaces/chat.interface";
 
-export default function ChatPanel(){
+export default function ChatPanel({chatHandler}: IChatPanelProp){
 
     const data_service = useContext(DataServiceContext)
     const api_service = useContext(httpRequestContext)
     const websocket_service = useContext(WebSocketServiceContext)
 
     const url = data_service.setUrl('chat_list')
-    
+
 
     const [loading, setLoading] = useState(true)
     const [chatsOn, setChatsOn] = useState<IChat[]>([])
     const [chatsOff, setChatsOff] = useState<IChat[]>([])
 
+
     useEffect(()=>{
         getChats()
         connectAdmin()
+        setLoading(false)
     },[])
+    
+
+
+
 
     const connectAdmin = () => {
         const ws_url = data_service.setUrlws("admin/")
         websocket_service.SetSocket(ws_url, WebsocketTypeEnum.Admin, getChats)
     }
-
 
 
     const getChats = async () => {
@@ -55,67 +63,78 @@ export default function ChatPanel(){
         let chats_on: IChat[] = chats.filter((chat) => chat.status === ChatStatusEnum.Open)
         let chats_off: IChat[] =  chats.filter((chat) => chat.status === ChatStatusEnum.Closed)
 
-        console.log(chats_on, chats_off)
-
+        console.log(chats_off, chats_on)
         setChatsOn(chats_on)
         setChatsOff(chats_off)
 
-        setLoading(false)
     }
 
+
+
+    // Channge this an put it on the individual sections
     if (loading){
         return <Loading/>
     }
+    
 
     return (
-        <section className="w-full h-full grid grid-rows-2 gap-4 p-9">
-
+ 
+           
+        
+        <section className="w-full h-full grid grid-rows-2 col-start-2 p-10 lg:p-0 lg:min-h-0">
+            {/* {chatOpen ?
+                <section className="fixed">
+                    <Chat key={chatData?.roomID} actual_user={UserTypeEnum.Admin} chat_data={chatData}/>
+                </section>
+            :null} */}
+            
             <section className="w-full h-full ">
                 
-                <section className=" w-full h-full border-8 p-4 flex flex-col">
+                <section className=" w-full h-full border-8 lg:border-4 p-4 flex flex-col bg-slate-500">
                     <h2 className="text-center text-white text-3xl">Conectados</h2>
-                    <section className="w-full h-full border-2 cursor-pointer border-red-600 overflow-scroll">
+                    <section className="w-full h-full border-2 bg-slate-700 overflow-scroll">
                         
-  
-                        {
-                            chatsOn.map((chat, index)=> (
-                                <div className="w-full h-12 border-2 flex items-center justify-center" key={index}>
-                                    <p className="text-white text-center">Cliente: {chat.client.name} Telefono: {chat.client.cellphone}</p>
-                                </div>
-                            ))
+
+                        {chatsOn.length === 0 ? (
+                                <p className="text-center text-red-500 text-2xl">No hay usuarios conectados</p>
+                            ) :(
+                                    chatsOn.map((chat, index)=> (
+                                        <div className="w-full h-12 flex cursor-pointer items-center justify-center shadow-[rgba(0,_0,_0,_0.24)_0px_3px_8px] bg-gray-700" key={index} onClick={()=>{chatHandler(chat)}}>
+                                            <p className="text-white text-center">Cliente: {chat.client.name} Telefono: {chat.client.cellphone}</p>
+                                        </div>
+                                    ))
+                                )
                         }
-           
+
                     </section>
                 </section>
 
             </section>
+            
             <section className="w-full h-full ">
                 
-                <section className=" w-full h-full border-8 p-4 flex flex-col">
-                    <h2 className="text-center text-white text-3xl">Desconectados</h2>
-                    <section className="w-full h-full border-2 border-red-600 overflow-scroll">
+                <section className=" w-full h-full border-8 lg:border-4 p-4 flex flex-col bg-slate-500">
+                    <h2 className="text-center text-white text-3xl">Guardados</h2>
+                    <section className="w-full h-full border-2 bg-slate-700 overflow-scroll">
                         
-  
-                        {
-                            chatsOff.map((chat, index)=> (
-                                <div className="w-full h-12 border-2 flex items-center justify-center" key={index}>
-                                    <p className="text-white text-center">Cliente: {chat.client.name} Telefono: {chat.client.cellphone}</p>
-                                </div>
-                            ))
+
+                        {chatsOff.length === 0 ? (
+                                <p className="text-center text-red-500 text-2xl"> No hay chats guardados</p>
+                            ) :(
+                                    chatsOff.map((chat, index)=> (
+                                        <div className="w-full h-12 flex cursor-pointer items-center justify-center shadow-[rgba(0,_0,_0,_0.24)_0px_3px_8px] bg-gray-700" key={index} onClick={()=>{chatHandler(chat)}}>
+                                            <p className="text-white text-center">Cliente: {chat.client.name} Telefono: {chat.client.cellphone}</p>
+                                        </div>
+                                    ))
+                                )
                         }
-                        
+
                     </section>
                 </section>
 
             </section>
             
         </section>
+
     )
-}
-
-function ChatRow({name, cellphone}: IClient){
-
-    <div className="w-full h-12 border-2 flex items-center justify-center">
-        <p className="text-white text-center">Cliente: {name} Telefono: {cellphone}</p>
-    </div>
 }