+++ Frontend/src/app/_components/admin/chat/chat.tsx
@@ -1,9 +1,0 @@
-export default function Chat(){
-
-
-    return (
-        <section className="w-full h-full bg-red-300 ">
-           
-        </section>
-    )
-}