+++ Frontend/src/app/_components/admin/chat/interfaces/chat.interface.ts
@@ -1,10 +1,15 @@
+import { IMessageData } from "@/app/_components/chat/interfaces/chat.interface"
+
+export interface IChatPanelProp{
+    chatHandler: (chat:IChat) => void
+}
+
 export interface IChat{
     created:string
     roomID: string
     status: string
     client: IClient
-    messages: IMessage[]
-
+    messages: IMessageData
     
 }
 
@@ -13,7 +18,3 @@ export interface IClient{
     cellphone: number
 }
 
-export interface IMessage{
-    user_type: string
-    text: string
-}