+++ Frontend/src/app/_components/admin/stats/stats.tsx
@@ -1,0 +1,7 @@
+export default function Stats(){
+
+    return (
+        <section className="w-full h-full bg-slate-500">
+        </section>
+    )
+}