+++ Frontend/src/app/_components/chat/enums/chat.enum.ts
@@ -1,5 +1,15 @@
-export enum messageType{
+export enum UserTypeEnum{
     AnonymousUser = 'AnonymousUser',
     Admin = 'Admin'
+
+}
+
+export enum MessageTypeEnum{
+    Chat = 0,
+    Notification = 1
+}
+export enum MessageOriginEnum{
+    Incoming = 'Incoming',
+    Outgoing = 'Outgoing'
 
 }