+++ Frontend/src/app/_components/chat/interfaces/chat.interface.ts
@@ -1,4 +1,26 @@
 
+
+export interface IChatProp{
+    actual_user: string
+    chat_data?: IChatData
+}
+
+export interface IChatData{
+    roomID: string
+    status: string
+    client_data: IClientData
+}
+
+export interface IClientData{
+    name: string
+    cellphone: number
+}
+
+export interface IMessageProp{
+    origin: string
+    message: IMessageData
+}
+
 export interface IMessage{
     type: number
     data: IMessageData
@@ -6,6 +28,7 @@ export interface IMessage{
 
 export interface IMessageData{
     user_type?: string
+    message_type?: number
     text: string
 }